+++ frontend/src/app/core/helpers/options-helper.ts
@@ -1,4 +1,4 @@
-import { ICheckOption } from '@core/interfaces/check-option';
+import { ICheckOption } from '@core/interfaces/check-option-interface';
 
 export const getDefaultOptions = (): ICheckOption[] => [
     { name: 'Yes', value: true },
