+++ frontend/src/app/core/interfaces/check-option-interface.ts
